requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dateutil==2.8.2
pyodbc==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
